Add dispute struct
code refactoring
